Add conventions
Cartographer Conventions is a new component that is part of
Cartographer. Adding the latest release of the conventions to the
exiting cartographer package. The version numbers for each component are
distinct and intended to drift. All resources are installed into the
cartographer-system namespace.

There are two user facing values:
- `ca_cert_data` allows users to trust additional ca certificates, which
  are commonly used inside corporate networks for internal image
  registries.
- `excluded_components` a list of components that should not be
  installed. The only valid values at the moment are `conventions` and
  `cartographer.

There are a handful of places where cartographer and cartographer
conventions have different behavior that we can iron out over time. They
include:

- image pull secrets - conventions defines the secret and injects the
  pull secret reference inside the bundle.
- aggregated roles - cartographer conventions can drop their inline
  roles now that the aggregated values are defined in package.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>