vendir: add cartographer-catalog
`vmware-tanzu/cartographer-catalog` is another project that we'd like to
leverage the packaging that we perform here in this repository, so,
include it in `vendir.yml`.

the idea behind `cartographer-catalog` is that we're able to pack into
that repository a set of reusable blueprints that projects like
vmware-tanzu/community-edition can benefit of, aside from the community
as a whole.

in this commit i'm adding the first iteration of
`vmware-tanzu/cartographer-catalog` which is a personal fork under
`cirocosta/ootb-tce` (temporary until the get the vmware-tanzu'
originated repository public).

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>