hack: point script at disting `./src` dirs
by letting the packaging that `hack/bundle.sh` performs take a directory
as an argument we're able to have the directories under `./src`
represent individual projects that we want to package up under the
`cartographer` umbrella.

now, `./hack/bundle.sh` is meant to be used as such:

	./hack/bundle.sh <directory_to_package>

such that to package cartographer, we'd do as such:

	./hack/bundle.sh ./src/cartographer

to enable that, each direct directory under `./src` carries a
`package-values.yaml` file whose whole goal is to provide
packaging-related configuration for the templating of carvel-based
packaging objects that `hack/bundle.sh` creates.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>